Added visualisation for visit counts
For each instruction it is counted how often it has been read.
Subsequently these counts are visualised via a heatmap palette.
Mapping to the limited colors in the heatmap is done by collapsing multiple counts to the same color in such a way to minimize the "area" of each bucket, where the area is the number of counts is represents multiplied by their range.