Minor layout change in Edit view
Moved display of available instructions to the right of the screen.
This to avoid it overlapping with exits of newly added challenges, which both have exits to the left of the board.
Obviously, we may still get a clash when there's a challenge with an exit at the top right, but we'll cross that bridge once we get to it.