Increased size of data store
It turns out that there are programs that terminate which can generate long sequences. In particular, I recently found a Busy Beaver for 7x7 that generates a sequence of length 74.

The data size is increased to 256 to accommodate this.
Furthermore, the data buffer is now cyclic so that the data storage can be used optimally.